test(install): expose test helpers
