+++ packages/ubuntu_desktop_installer/test/install/install_model_test.dart
@@ -1,16 +1,14 @@
 import 'dart:async';
 
 import 'package:flutter_test/flutter_test.dart';
-import 'package:mockito/annotations.dart';
 import 'package:mockito/mockito.dart';
 import 'package:subiquity_client/subiquity_client.dart';
 import 'package:subiquity_test/subiquity_test.dart';
 import 'package:ubuntu_desktop_installer/pages/install/install_model.dart';
 import 'package:ubuntu_desktop_installer/services.dart';
 
-import 'install_model_test.mocks.dart';
+import 'test_install.dart';
 
-@GenerateMocks([JournalService, ProductService])
 void main() async {
   test('product info', () {
     final product = MockProductService();
