+++ packages/ubuntu_desktop_installer/test/install/test_install.mocks.dart
@@ -1,5 +1,5 @@
 // Mocks generated by Mockito 5.4.0 from annotations
-// in ubuntu_desktop_installer/test/install/install_page_test.dart.
+// in ubuntu_desktop_installer/test/install/test_install.dart.
 // Do not manually edit this file.
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
@@ -42,6 +42,53 @@ class _FakeInstallationEvent_1 extends _i1.SmartFake
           parent,
           parentInvocation,
         );
+}
+
+/// A class which mocks [JournalService].
+///
+/// See the documentation for Mockito's code generation for more information.
+class MockJournalService extends _i1.Mock implements _i2.JournalService {
+  MockJournalService() {
+    _i1.throwOnMissingStub(this);
+  }
+
+  @override
+  _i4.Stream<String> start(
+    List<String>? ids, {
+    _i2.JournalOutput? output = _i2.JournalOutput.short,
+  }) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #start,
+          [ids],
+          {#output: output},
+        ),
+        returnValue: _i4.Stream<String>.empty(),
+      ) as _i4.Stream<String>);
+}
+
+/// A class which mocks [ProductService].
+///
+/// See the documentation for Mockito's code generation for more information.
+class MockProductService extends _i1.Mock implements _i2.ProductService {
+  MockProductService() {
+    _i1.throwOnMissingStub(this);
+  }
+
+  @override
+  _i2.ProductInfo getProductInfo() => (super.noSuchMethod(
+        Invocation.method(
+          #getProductInfo,
+          [],
+        ),
+        returnValue: _FakeProductInfo_0(
+          this,
+          Invocation.method(
+            #getProductInfo,
+            [],
+          ),
+        ),
+      ) as _i2.ProductInfo);
 }
 
 /// A class which mocks [InstallModel].
